Part 4 exercise 4.21

+++ controllers/blogs.js
@@ -29,11 +29,11 @@ blogsRouter.get("/", async (request, response) => {
 blogsRouter.post("/", async (request, response, next) => {
   const body = request.body;
 
-  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
   if (!decodedToken.id) {
-    return response.status(401).json({ error: 'token missing or invalid' })
+    return response.status(401).json({ error: "token missing or invalid" });
   }
-  const user = await User.findById(decodedToken.id)
+  const user = await User.findById(decodedToken.id);
 
   // const users = await User.find({});
   // const id = users[0].id;
@@ -60,8 +60,26 @@ blogsRouter.post("/", async (request, response, next) => {
 });
 
 blogsRouter.delete("/:id", async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id);
-  response.status(204).end();
+  const body = request.body;
+
+  if (!request.token) {
+    return response.status(401).json({ error: "token missing" });
+  }
+  
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: "invalid token" });
+  }
+
+  const user = await User.findById(decodedToken.id);
+  const blog = await Blog.findById(request.params.id);
+
+  if (blog.user.toString() === decodedToken.id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id);
+    response.status(204).end();
+  } else {
+    return response.status(401).json({ error: "wrong user" });
+  }
 });
 
 blogsRouter.get("/:id", async (request, response) => {
